pyproject.toml
[project]
name = "seo-reporting-system"
version = "0.1.0"
description = "Monthly SEO report generator (spec-compliant)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "jsonschema>=4.22.0",
  "duckdb>=1.0.0",
  "jinja2>=3.1.3",
  "requests>=2.32.0",
  "pyyaml>=6.0.1",
  "tenacity>=8.3.0",
  "python-dotenv>=1.0.1",
  "google-auth>=2.30.0",
]

[project.scripts]
seo-report = "app.cli:app"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
